Conflicts:
	doc/Makefile
	doc/stdlib/index.rst

+++ base/git.jl
@@ -3,8 +3,6 @@ module Git
 # some utility functions for working with git repos
 #
 
-using Base
-
 dir() = readchomp(`git rev-parse --git-dir`)
 modules(args::Cmd) = readchomp(`git config -f .gitmodules $args`)
 different(verA::String, verB::String, path::String) =
@@ -17,11 +15,12 @@ dirty(paths) = !success(`git diff --quiet HEAD -- $paths`)
 staged(paths) = !success(`git diff --quiet --cached -- $paths`)
 unstaged(paths) = !success(`git diff --quiet -- $paths`)
 
-attached() = success(`git symbolic-ref -q HEAD` > "/dev/null")
+attached() = success(`git symbolic-ref -q HEAD` > SpawnNullStream())
 branch() = readchomp(`git rev-parse --symbolic-full-name --abbrev-ref HEAD`)
+head() = readchomp(`git rev-parse HEAD`)
 
-function each_version()
-    git_dir = abs_path(dir())
+function each_tagged_version()
+    git_dir = abspath(dir())
     @task for line in each_line(`git --git-dir=$git_dir show-ref --tags`)
         m = match(r"^([0-9a-f]{40}) refs/tags/(v\S+)$", line)
         if m != nothing && ismatch(Base.VERSION_REGEX, m.captures[2])
@@ -29,12 +28,13 @@ function each_version()
         end
     end
 end
-each_version(dir::String) = cd(each_version,dir)
+each_tagged_version(dir::String) = cd(each_tagged_version,dir)
 
 function each_submodule(f::Function, recursive::Bool, dir::ByteString)
-    cmd = `git submodule foreach --quiet 'echo "$name\t$path\t$sha1"'`
+    cmd = `git submodule foreach --quiet 'echo "$name $path $sha1"'`
     for line in each_line(cmd)
-        name, path, sha1 = match(r"^(.*)\t(.*)\t([0-9a-f]{40})$", line).captures
+        isempty(line) && break # FIXME: temporary work-around #2089
+        name, path, sha1 = match(r"^(.*) (.*) ([0-9a-f]{40})$", line).captures
         cd(dir) do
             f(name, path, sha1)
         end
@@ -49,7 +49,7 @@ function each_submodule(f::Function, recursive::Bool, dir::ByteString)
         end
     end
 end
-each_submodule(f::Function, r::Bool) = each_submodule(f, r, cwd())
+each_submodule(f::Function, r::Bool) = each_submodule(f, r, pwd())
 
 function read_config(file::String)
     cfg = Dict()
@@ -84,8 +84,10 @@ function write_config(file::String, cfg::Dict)
             run(`git config -f $tmp $key $val`)
         end
     end
-    open(file,"w") do io
-        print(io,readall(tmp))
+    if isfile(tmp)
+        open(file,"w") do io
+            print(io,readall(tmp))
+        end
     end
 end
 
@@ -129,6 +131,21 @@ function merge_configs(Bc::Dict, Lc::Dict, Rc::Dict)
         end
     end
     return cfg, conflicts, deleted
+end
+
+const GITHUB_REGEX = r"^(?:git@|git://|https://(?:[\w\.\+\-]+@)?)github.com[:/](.*)$"i
+
+# setup a repo's push URL intelligently
+
+function autoconfig_pushurl()
+    url = readchomp(`git config remote.origin.url`)
+    m = match(GITHUB_REGEX,url)
+    if m != nothing
+        pushurl = "git@github.com:$(m.captures[1])"
+        if pushurl != url
+            run(`git config remote.origin.pushurl $pushurl`)
+        end
+    end
 end
 
 end # module
